authorizaion for edit and update with gate

+++ app/Http/Controllers/QuestionsController.php
@@ -61,7 +61,12 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        if(\Gate::denies('update-question',$question))
+        {
+            abort(403,"Access");
+        }
         return view('questions.edit',compact('question'));
+  
     }
 
     /**
@@ -73,6 +78,10 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if(\Gate::denies('update-question',$question))
+        {
+            abort(403,"Access Denided");
+        }
         $question->update($request->only('title','body'));
         return redirect()->route('questions.index')->with('success','Your Question Updated');
     }
@@ -85,6 +94,10 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if(\Gate::denies('delete-question',$question))
+        {
+            abort(403,"Access Denided");
+        }
         $question->delete();
         return redirect()->route('questions.index')->with('success','Your Question Deleted');
     }
